+++ index.js
@@ -34,14 +34,29 @@ function forwarded (req) {
   }
 
   // simple header parsing
-  var proxyAddrs = (req.headers['x-forwarded-for'] || '')
-    .trim()
-    .split(TOKEN_LIST_REGEXP)
-    .filter(Boolean)
-    .reverse()
+  var proxyAddrs = parse(req.headers['x-forwarded-for'] || '')
   var socketAddr = req.connection.remoteAddress
   var addrs = [socketAddr].concat(proxyAddrs)
 
   // return all addresses
   return addrs
+}
+
+/**
+ * Parse the X-Forwarded-For header.
+ *
+ * @param {string} header
+ * @private
+ */
+
+function parse (header) {
+  if (!header) {
+    return []
+  }
+
+  return header
+    .trim()
+    .split(TOKEN_LIST_REGEXP)
+    .filter(Boolean)
+    .reverse()
 }